request todos by connected user

+++ app/src/infrastructure/repositories/todo/todo.repository.ts
@@ -1,15 +1,19 @@
-import { Injectable } from "@nestjs/common";
+import { Inject, Injectable, Scope } from "@nestjs/common";
+import { REQUEST } from "@nestjs/core";
 import { InjectRepository } from "@nestjs/typeorm";
+import type { Request } from "express";
 import { TodoModel } from "src/domain/model/todo";
+import type { UserModel } from "src/domain/model/user";
 import type { TodoRepository } from "src/domain/repositories/todo.interface";
 import { Todo } from "src/infrastructure/entities/todo.entity";
 import type { Repository } from "typeorm";
 
-@Injectable()
+@Injectable({ scope: Scope.REQUEST })
 export class DatabaseTodoRepository implements TodoRepository {
   constructor(
     @InjectRepository(Todo)
     private readonly todoEntityRepository: Repository<Todo>,
+    @Inject(REQUEST) private request: Request,
   ) {}
 
   async insert(todo: Pick<TodoModel, "content" | "isDone">): Promise<void> {
@@ -22,7 +26,7 @@ export class DatabaseTodoRepository implements TodoRepository {
         content: todo.content,
         createdAt: new Date(),
         updatedAt: new Date(),
-        creator: { id: 2, mail: "user@mail.fr", username: "user" },
+        creator: this.request.user,
       })
       .execute();
   }
@@ -31,6 +35,9 @@ export class DatabaseTodoRepository implements TodoRepository {
     const todosEntity = await this.todoEntityRepository
       .createQueryBuilder("todo")
       .leftJoinAndSelect("todo.creator", "user")
+      .where("user.id = :userId", {
+        userId: (this.request.user as Pick<UserModel, "id">).id,
+      })
       .getMany();
     const todos = todosEntity.map((todoEntity) => this.toTodo(todoEntity));
     return todos;
@@ -41,6 +48,9 @@ export class DatabaseTodoRepository implements TodoRepository {
       .createQueryBuilder("todo")
       .leftJoinAndSelect("todo.creator", "user")
       .where("todo.id = :id", { id })
+      .andWhere("user.id = :userId", {
+        userId: (this.request.user as Pick<UserModel, "id">).id,
+      })
       .getOne();
     if (!todoEntity) {
       return null;
@@ -53,6 +63,9 @@ export class DatabaseTodoRepository implements TodoRepository {
       .createQueryBuilder("todo")
       .leftJoinAndSelect("todo.creator", "user")
       .where("todo.content LIKE :search", { search: `%${search}%` })
+      .andWhere("user.id = :userId", {
+        userId: (this.request.user as Pick<UserModel, "id">).id,
+      })
       .getMany();
     const todos = todoEntities.map((todoEntity) => this.toTodo(todoEntity));
     return todos;
@@ -63,6 +76,9 @@ export class DatabaseTodoRepository implements TodoRepository {
       .createQueryBuilder("todo")
       .delete()
       .where("todo.id = :id", { id })
+      .where("user.id = :userId", {
+        userId: (this.request.user as Pick<UserModel, "id">).id,
+      })
       .execute();
   }
 
@@ -79,6 +95,9 @@ export class DatabaseTodoRepository implements TodoRepository {
         creator: todo.creator,
       })
       .where("todo.id = :id", { id })
+      .where("user.id = :userId", {
+        userId: (this.request.user as Pick<UserModel, "id">).id,
+      })
       .execute();
   }
